Feat: 로그들을 가져오는 엔드포인트 구현 완료

+++ video/src/main/java/toykiwi/sanityCheck/LogsReqDto.java
@@ -6,5 +6,5 @@ import lombok.ToString;
 @Data
 @ToString
 public class LogsReqDto {
-    private Long lineLength = 10L;
+    private int lineLength = 10;
 }
